Renamed memory.* files to avoid conflicts
