+++ include/rmm/detail/memory_manager.hpp
@@ -31,7 +31,7 @@
 #include <set>
 #include <mutex>
 
-#include "rmm/memory.h"
+#include "rmm/rmm_api.h"
 #include "rmm/detail/cnmem.h"
 
 /** ---------------------------------------------------------------------------*
